backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
aniso8601==10.0.1
annotated-types==0.7.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clerk_backend_sdk==1.1.1
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-APScheduler==1.13.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.22.0
